pyproject.toml
[project]
name = "scaffold"
version = "0.1.0"
description = "Estrutura básica do projeto Django com Poetry"
authors = [
    {name = "JL Gatz",email = "elan.rgb@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "djantic (>=0.7.0,<0.8.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{'include' = 'equipamentos'}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.13"
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
taskipy = "^1.14.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations', 'manage.py', 'settings.py']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'python manage.py runserver'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
make = 'python manage.py makemigrations'
migrate = 'python manage.py migrate'

